refactored some code
