+++ app/proguard-rules-release-preprod.pro
